Fixed broken link to MS UWP page
In the .NET Native section, it links to a Microsoft page explaining UWP. The link is broken. The initial URL is:

https://msdn.microsoft.com/library/windows/apps/dn726767.aspx

which redirects to: https://www.visualstudio.com/docs/whats-a-uwp

which yields a page not found.

I replaced the link URL with https://msdn.microsoft.com/en-us/windows/uwp/layout/design-and-ui-intro which was the second highest result in Google for "whats a uwp".

(Note that the first Google result is also a broken link: https://msdn.microsoft.com/en-us/windows/uwp/get-started/whats-a-uwp. You might want to fix the issue with that article being gone everywhere.)